+++ src/index.ts
@@ -1,2 +1,3 @@
 export { ERC20Proof } from "./erc20"
 export { MiniMeProof } from "./minime"
+export { EthProofs } from "./common"