+++ test/integration/erc20.ts
@@ -1,6 +1,6 @@
 import "mocha" // using @types/mocha
 import { expect } from "chai"
-import { ERC20Proof } from "../../src/index"
+import { ERC20Proof, EthProofs } from "../../src/index"
 import { provider } from "../util"
 import { addCompletionHooks } from "../mocha-hooks"
 
@@ -25,7 +25,7 @@ describe('ERC20 Storage Proofs', () => {
         ]
 
         for (let item of data) {
-            expect(ERC20Proof.getHolderBalanceSlot(item.addr, item.idx)).to.eq(item.output)
+            expect(EthProofs.getMapSlot(item.addr, item.idx)).to.eq(item.output)
         }
     })
 
@@ -33,7 +33,7 @@ describe('ERC20 Storage Proofs', () => {
         const holderAddr = "0x1062a747393198f70f71ec65a582423dba7e5ab3"
 
         blockNumber = await provider.getBlockNumber()
-        const balanceSlot = ERC20Proof.getHolderBalanceSlot(holderAddr, BALANCE_MAPPING_SLOT)
+        const balanceSlot = EthProofs.getMapSlot(holderAddr, BALANCE_MAPPING_SLOT)
         const result = await ERC20Proof.get(TOKEN_ADDRESS, [balanceSlot], blockNumber, provider)
 
         expect(result.proof).to.be.ok
@@ -60,7 +60,7 @@ describe('ERC20 Storage Proofs', () => {
     it('Should proof that a value does exist', async () => {
         const holderAddress = "0x1062a747393198f70f71ec65a582423dba7e5ab3"
 
-        const balanceSlot = ERC20Proof.getHolderBalanceSlot(holderAddress, BALANCE_MAPPING_SLOT)
+        const balanceSlot = EthProofs.getMapSlot(holderAddress, BALANCE_MAPPING_SLOT)
         const storageKeys = [balanceSlot]
         const { block, proof } = await ERC20Proof.get(TOKEN_ADDRESS, storageKeys, "latest", provider)
         expect(proof.storageProof[0].value).to.not.eq("0x0")
@@ -76,7 +76,7 @@ describe('ERC20 Storage Proofs', () => {
         const holderAddress = "0x0010000000000000000000000000000000000000"
         const tokenAddress = "0x6b175474e89094c44da98b954eedeac495271d0f"
 
-        const balanceSlot = ERC20Proof.getHolderBalanceSlot(holderAddress, 100)
+        const balanceSlot = EthProofs.getMapSlot(holderAddress, 100)
         const storageKeys = [balanceSlot]
 
         const { proof, block } = await ERC20Proof.get(tokenAddress, storageKeys, "latest", provider)
@@ -94,7 +94,7 @@ describe('ERC20 Storage Proofs', () => {
         const tokenAddress = "0x6b175474e89094c44da98b954eedeac495271d0f"
         const unrealBalanceMappingPosition = 100
 
-        const balanceSlot = ERC20Proof.getHolderBalanceSlot(holderAddress, unrealBalanceMappingPosition)
+        const balanceSlot = EthProofs.getMapSlot(holderAddress, unrealBalanceMappingPosition)
         const storageKeys = [balanceSlot]
 
         {